Put both warnings into exceptions submodule
