+++ docs/source/api.rst
@@ -2,6 +2,7 @@ stdatamodels API
 ================
 
 .. automodapi:: stdatamodels
+.. automodapi:: stdatamodels.exceptions
 .. automodapi:: stdatamodels.asdf_in_fits
 .. automodapi:: stdatamodels.jwst.datamodels
