+++ src/stdatamodels/__init__.py
@@ -1,9 +1,8 @@
 from .model_base import DataModel
-from .validate import ValidationWarning
 from . import _version
 
 
-__all__ = ["DataModel", "ValidationWarning", "__version__"]
+__all__ = ["DataModel", "__version__"]
 
 
 __version__ = _version.version
