+++ src/stdatamodels/jwst/datamodels/__init__.py
@@ -112,12 +112,11 @@ from .wcs_ref_models import (
     WaveCorrModel,
 )
 from .wfssbkg import WfssBkgModel
-from .util import open, NoTypeWarning  # noqa: A004
+from .util import open  # noqa: A004
 
 
 __all__ = [
     "open",
-    "NoTypeWarning",
     "JwstDataModel",
     "ABVegaOffsetModel",
     "AmiLgModel",
@@ -252,7 +251,7 @@ __all__ = [
 ]
 
 
-_all_models = __all__[2:]
+_all_models = __all__[1:]
 _local_dict = locals()
 _defined_models = {k: _local_dict[k] for k in _all_models}
