+++ src/stdatamodels/jwst/datamodels/reference.py
@@ -1,6 +1,6 @@
 import warnings
 
-from stdatamodels.validate import ValidationWarning
+from stdatamodels.exceptions import ValidationWarning
 from stdatamodels.dynamicdq import dynamic_mask
 
 from .model_base import JwstDataModel
