+++ src/stdatamodels/jwst/datamodels/tsophot.py
@@ -2,7 +2,7 @@ import warnings
 import sys
 import traceback
 
-from stdatamodels.validate import ValidationWarning
+from stdatamodels.exceptions import ValidationWarning
 
 from .reference import ReferenceFileModel
