+++ src/stdatamodels/jwst/datamodels/util.py
@@ -14,24 +14,14 @@ import numpy as np
 from astropy.io import fits
 from stdatamodels import filetype
 from stdatamodels.model_base import _FileReference
+from stdatamodels.exceptions import NoTypeWarning
 
 
-__all__ = ["open", "NoTypeWarning", "is_association"]
+__all__ = ["open", "is_association"]
 
 
 log = logging.getLogger(__name__)
 log.addHandler(logging.NullHandler())
-
-
-class NoTypeWarning(Warning):
-    """
-    Warning to raise when opening a file that lacks a model type.
-
-    For ASDF files this means the meta.model_type keyword is missing.
-    For FITS files this means the DATAMODL keyword is missing from the primary hdulist.
-    """
-
-    pass
 
 
 def open(init=None, guess=True, memmap=False, **kwargs):  # noqa: A001
