+++ src/stdatamodels/jwst/datamodels/validate.py
@@ -1,4 +1,4 @@
-from stdatamodels.validate import ValidationWarning
+from stdatamodels.exceptions import ValidationWarning
 
 
 # Keep this here for backwards compatibility
